Refactor EventPublishingException to be uncheked exception
