--- src/main/java/org/zalando/nakadi/exceptions/EventPublishingException.java
@@ -1,20 +1,0 @@
-package org.zalando.nakadi.exceptions;
-
-import org.zalando.problem.MoreStatus;
-
-import javax.ws.rs.core.Response;
-
-public class EventPublishingException extends NakadiException {
-    public EventPublishingException(final String message, final Exception e) {
-        super(message, e);
-    }
-
-    public EventPublishingException(final String message) {
-        super(message);
-    }
-
-    @Override
-    protected Response.StatusType getStatus() {
-        return MoreStatus.MULTI_STATUS;
-    }
-}
